move vat starting state into kernel
